Add API endpoint to get an author by ID
This isn't meant to be much apart from giving me an idea of how Flask works. This adds an API endpoint to get the author by ID.
Also, random idea came to mind just now: If we show the author's profile, we should have a similar teachers section that shows other professors who teach courses similar to the current authors own.

This relationship could be described like this:

CurrentAuthor => Teaches => Courses => Tagged With => Tags => Related Courses => Coursess => Authors

Yow, the thought that that might be possible is freaking insane haha. Imagine the SQL version of that. :(